refactor: move query-client hook
